Move solution stat collection into onSolutionSubmission (#144)
This PR moves the collection of solution specific details such as solver, objective value, utility, etc from the first `onTrade` handler to a specific `onSolutionSubmission` handler. Main motivation for this is that we want to record global stats such as fees, utility, volume and for that it's simpler to use a designated top level handler (it would be possible using todays setup but involve much more computation e.g. when deriving the utility of a certain traded). On the contrary, utility in OWL is part of the SolutionSubmission event already.

In this PR I'm mainly changing the status quo. In the next PR I will be adding a global Stat entity that also gets populated inside the onSolutionSubmission method.

It has the nice side effect that we no longer need to mock the call to `latestSolution` in our unit tests.

### Test Plan

CI